+++ sketch/sound.js
@@ -1,6 +1,7 @@
 new p5();
 
 let backgroundColor;
+let opacity = 10;
 
 let song;
 let amp;
@@ -8,6 +9,12 @@ let vol;
 let volHistory = [];
 
 let circle = true;
+
+let fft;
+let bands = Math.pow(2, 7);
+let smoothing = 0.999;
+
+let w;
 
 let button;
 
@@ -20,14 +27,18 @@ function setup() {
 
 	let canvas = createCanvas(
 		window.innerWidth, //* 7/10,
+		// 700,
 		window.innerHeight // * 7/10
 		);
 
-	backgroundColor = color(200);
+	backgroundColor = color(240, opacity);
 
 	song.play();
 
 	amp = new p5.Amplitude();
+	fft = new p5.FFT(smoothing, bands);
+
+	w = width/bands;
 
 	// button = createButton('Toggle');
 	// button.mousePressed(toggleSong);
@@ -35,18 +46,23 @@ function setup() {
 
 function draw() {
 	background(backgroundColor);
-	
-	vol = amp.getLevel();
-	volHistory.push(vol);
-	limitVolHistory();
 
-	displayGraph();
+	// handleAmpInput();
+	// displayAmpGraph();
+
+	handleFFT();
 }
 
 function windowResized() {
 	resizeCanvas(
 		window.innerWidth,
 		window.innerHeight);
+}
+
+function handleAmpInput() {
+	vol = amp.getLevel();
+	volHistory.push(vol);
+	limitVolHistory();
 }
 
 function toggleSong() {
@@ -58,19 +74,14 @@ function toggleSong() {
 }
 
 function limitVolHistory() {
-	let maxSize;
-	// if (circle) {
-	// 	maxSize = 360;
-	// } else {
-		maxSize = width;
-	// }
+	let maxSize = width;
 
 	if (volHistory.length > maxSize) {
 		volHistory.splice(0, volHistory.length - maxSize);
 	}
 }
 
-function displayGraph() {
+function displayAmpGraph() {
 	if (circle) {
 		angleMode(RADIANS);
 		drawCircleGraph(color(220));		
@@ -122,3 +133,57 @@ function displayMouth() {
 	noStroke();
 	ellipse(width/2, height/2, width*8/10, vol*1000);
 }
+
+function handleFFT() {
+	let spectrum = fft.analyze();
+	// console.log(spectrum);
+	drawFrequencySpectrum(spectrum);
+}
+
+function drawFrequencySpectrum(spectrum) {
+	noFill();
+	strokeWeight(1);
+	
+	let y = 0;
+	// let greyValue = 0;
+	let hue = 0;
+	let barAlpha = 70;
+	
+	let peakAlpha = 90;
+	let peakColor = color(200, peakAlpha);
+	
+	beginShape();
+	// Draw frequency bar
+	for (let i=0; i<spectrum.length; i++) {
+		y = map(spectrum[i], 0, 255, height, 0);
+		
+		// greyValue = map(spectrum[i], 0, 255, 50, 245);
+		hue = map(spectrum[i], 0, 255, 170, 10);
+		colorMode(HSB);
+		
+		// Extra projection lines
+		noFill();
+		stroke(hue, 20, 100, 10);
+		strokeWeight(1);
+		// line(0, height, i*w, y+1); // +1 to remove interference with vertices;
+		line(width/2, height/2, i*w +w/2, y); // 
+		
+		// Frequency bars
+		let frBarColor = color(hue, 255, 100, barAlpha);
+		strokeWeight(1);
+		stroke(frBarColor);
+		fill(frBarColor);
+		rect(i*w, y, w*9.5/10, height-y, 30, 30, 0, 0);
+		
+		// Peak vertices
+		vertex(i*w + w/2, y);
+
+	}
+
+	// Draw peak
+	noFill();
+	colorMode(RGB);
+	stroke(peakColor);
+	strokeWeight(w*2/3);
+	endShape();
+}